requirements.txt
addict==2.4.0
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
bytecode==0.15.1
certifi @ file:///croot/certifi_1665076670883/work/certifi
charset-normalizer==2.1.1
clip @ git+https://github.com/openai/CLIP.git@d50d76daa670286dd6cacf3bcd80b5e4823fc8e1
contourpy==1.0.5
cycler==0.11.0
datasets==2.14.4
dill==0.3.7
einops==0.5.0
faiss-gpu==1.7.2
filelock==3.12.2
fonttools==4.37.4
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.1.1
h5py==3.9.0
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==7.0.0
joblib==1.3.2
jsonpatch==1.32
jsonpointer==2.3
kiwisolver==1.4.4
markdown-it-py==3.0.0
matplotlib==3.6.1
mdurl==0.1.2
mmcv==2.1.0
mmengine==0.10.1
multidict==6.0.4
multiprocess==0.70.15
networkx==2.8.7
numpy==1.23.4
opencv-python==4.8.1.78
packaging==21.3
pandas==2.0.3
Pillow==9.2.0
platformdirs==4.1.0
pyarrow==13.0.0
pycocotools==2.0.5
Pygments==2.17.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
pyzmq==24.0.1
randaugment==1.0.2
regex==2022.9.13
requests==2.28.1
rich==13.7.0
safetensors==0.3.3
scikit-learn==1.3.2
scipy==1.9.2
six==1.16.0
termcolor==2.4.0
threadpoolctl==3.2.0
tokenizers==0.13.3
tomli==2.0.1
torch==1.12.1+cu116
torchaudio==0.12.1+cu116
torchnet==0.0.4
torchvision==0.13.1+cu116
tornado==6.2
tqdm==4.64.1
transformers==4.32.0
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.12
visdom==0.2.1
wcwidth==0.2.5
websocket-client==1.4.1
xxhash==3.3.0
yapf==0.40.2
yarl==1.9.2
zipp==3.17.0
